LayoutTests/accessibility/ios-simulator/accessibility-events-setting-expected.txt

Click
This tests that dispatching accessibility events is controled by iOS settings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS receivedAXEvent is true
PASS receivedAXEvent is false
PASS receivedAXEvent is true
PASS successfullyParsed is true

TEST COMPLETE
